pyproject.toml
[tool.poetry]
name = "generate-commentary"
version = "0.1.0"
description = ""
authors = ["Tanaka <tsunehiko@fuji.waseda.jp>"]

[tool.poetry.dependencies]
python = "^3.6.9"
future = "^0.18.2"
Click = "^7.1.2"
tqdm = "^4.46.1"
pandas = "^1.0.4"
sklearn = "^0.0"
numpy = "1.16.4"
tensorboardx = "^2.0"
pygments = "^2.6.1"
optuna = "^1.5.0"
plotly = "^4.8.2"
matplotlib = "^3.2.2"
webvtt-py = "^0.4.5"
timecode = "^1.2.0"
decord = "^0.4.0"
PyTurboJPEG = "^1.4.1"
torch = "^1.6.0"
scenedetect = "^0.5.3"
deepsegment = "^2.3.1"
tensorflow = "^2.3.1"
fastpunct = "^1.0.2"
nltk = "^3.5"
scipy = "^1.5.3"
opencv-python = "^4.4.0"
mmaction = "^0.5.0"
keras = "^2.4.3"
torchtext = "^0.7.0"
spacy = "^2.3.5"
torchvision = "^0.8.2"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
